fixed sync join and separated servers
